refactor(alchemy-signer): refactor the alchemy signer with base classes
