+++ packages/alchemy/src/signer/client/index.ts
@@ -1,26 +1,16 @@
-import {
-  ConnectionConfigSchema,
-  type ConnectionConfig,
-} from "@alchemy/aa-core";
-import { TurnkeyClient, getWebAuthnAttestation } from "@turnkey/http";
+import { ConnectionConfigSchema } from "@alchemy/aa-core";
+import { getWebAuthnAttestation } from "@turnkey/http";
 import { IframeStamper } from "@turnkey/iframe-stamper";
 import { WebauthnStamper } from "@turnkey/webauthn-stamper";
-import EventEmitter from "eventemitter3";
-import { type Hex } from "viem";
 import { z } from "zod";
-import { NotAuthenticatedError } from "../errors.js";
 import { base64UrlEncode } from "../utils/base64UrlEncode.js";
 import { generateRandomBuffer } from "../utils/generateRandomBuffer.js";
+import { BaseSignerClient } from "./base.js";
 import type {
-  AlchemySignerClientEvent,
-  AlchemySignerClientEvents,
   CreateAccountParams,
   CredentialCreationOptionOverrides,
   EmailAuthParams,
   ExportWalletParams,
-  SignerBody,
-  SignerResponse,
-  SignerRoutes,
   User,
 } from "./types";
 
@@ -45,71 +35,34 @@ export type AlchemySignerClientParams = z.input<
  * A lower level client used by the AlchemySigner used to communicate with
  * Alchemy's signer service.
  */
-export class AlchemySignerClient {
-  private _user: User | undefined;
-  private connectionConfig: ConnectionConfig;
+export class AlchemySignerWebClient extends BaseSignerClient<ExportWalletParams> {
   private iframeStamper: IframeStamper;
   private webauthnStamper: WebauthnStamper;
-  private turnkeyClient: TurnkeyClient;
-  private eventEmitter: EventEmitter<AlchemySignerClientEvents>;
-  private rootOrg: string;
   iframeContainerId: string;
 
   constructor(params: AlchemySignerClientParams) {
     const { connection, iframeConfig, rpId, rootOrgId } =
       AlchemySignerClientParamsSchema.parse(params);
-    this.connectionConfig = connection;
-    this.eventEmitter = new EventEmitter<AlchemySignerClientEvents>();
 
-    this.iframeStamper = new IframeStamper({
+    const iframeStamper = new IframeStamper({
       iframeElementId: iframeConfig.iframeElementId,
       iframeUrl: "https://auth.turnkey.com",
       iframeContainer: document.getElementById(iframeConfig.iframeContainerId),
     });
 
-    this.iframeContainerId = iframeConfig.iframeContainerId;
+    super({
+      connection,
+      rootOrgId,
+      stamper: iframeStamper,
+    });
 
-    this.turnkeyClient = new TurnkeyClient(
-      { baseUrl: "https://api.turnkey.com" },
-      this.iframeStamper
-    );
+    this.iframeStamper = iframeStamper;
+    this.iframeContainerId = iframeConfig.iframeContainerId;
 
     this.webauthnStamper = new WebauthnStamper({
       rpId: rpId ?? window.location.hostname,
     });
-
-    this.rootOrg = rootOrgId;
   }
-
-  private get user() {
-    return this._user;
-  }
-
-  private set user(user: User | undefined) {
-    if (user && !this._user) {
-      this.eventEmitter.emit("connected", user);
-    } else if (!user && this._user) {
-      this.eventEmitter.emit("disconnected");
-    }
-
-    this._user = user;
-  }
-
-  /**
-   * Listen to events emitted by the client
-   *
-   * @param event the event you want to listen to
-   * @param listener the callback function to execute when an event is fired
-   * @returns a function that will remove the listener when called
-   */
-  on = <E extends AlchemySignerClientEvent>(
-    event: E,
-    listener: AlchemySignerClientEvents[E]
-  ) => {
-    this.eventEmitter.on(event, listener as any);
-
-    return () => this.eventEmitter.removeListener(event, listener as any);
-  };
 
   public createAccount = async (params: CreateAccountParams) => {
     this.eventEmitter.emit("authenticating");
@@ -189,40 +142,6 @@ export class AlchemySignerClient {
     return user;
   };
 
-  public whoami = async (orgId = this.user?.orgId): Promise<User> => {
-    if (this.user) {
-      return this.user;
-    }
-
-    if (!orgId) {
-      throw new Error("No orgId provided");
-    }
-
-    const stampedRequest = await this.turnkeyClient.stampGetWhoami({
-      organizationId: orgId,
-    });
-
-    const user = await this.request("/v1/whoami", {
-      stampedRequest,
-    });
-
-    const credentialId = (() => {
-      try {
-        return JSON.parse(stampedRequest?.stamp.stampHeaderValue)
-          .credentialId as string;
-      } catch (e) {
-        return undefined;
-      }
-    })();
-
-    this.user = {
-      ...user,
-      credentialId,
-    };
-
-    return this.user;
-  };
-
   public lookupUserWithPasskey = async (user: User | undefined = undefined) => {
     this.eventEmitter.emit("authenticating");
     await this.initWebauthnStamper(user);
@@ -238,42 +157,6 @@ export class AlchemySignerClient {
     return result;
   };
 
-  public lookupUserByEmail = async (email: string) => {
-    return this.request("/v1/lookup", { email });
-  };
-
-  /**
-   * This will sign a message with the user's private key, without doing any transformations on the message.
-   * For SignMessage or SignTypedData, the caller should hash the message before calling this method and pass
-   * that result here.
-   *
-   * @param msg the hex representation of the bytes to sign
-   * @returns the signature over the raw hex
-   */
-  public signRawMessage = async (msg: Hex) => {
-    if (!this.user) {
-      throw new NotAuthenticatedError();
-    }
-
-    const stampedRequest = await this.turnkeyClient.stampSignRawPayload({
-      organizationId: this.user.orgId,
-      type: "ACTIVITY_TYPE_SIGN_RAW_PAYLOAD_V2",
-      timestampMs: Date.now().toString(),
-      parameters: {
-        encoding: "PAYLOAD_ENCODING_HEXADECIMAL",
-        hashFunction: "HASH_FUNCTION_NO_OP",
-        payload: msg,
-        signWith: this.user.address,
-      },
-    });
-
-    const { signature } = await this.request("/v1/sign-payload", {
-      stampedRequest,
-    });
-
-    return signature;
-  };
-
   public exportWallet = async ({
     iframeContainerId,
     iframeElementId = "turnkey-export-iframe",
@@ -286,160 +169,16 @@ export class AlchemySignerClient {
     await exportWalletIframeStamper.init();
 
     if (this.turnkeyClient.stamper === this.iframeStamper) {
-      return this.exportAsSeedPhrase(exportWalletIframeStamper);
+      return this.exportWalletInner({
+        exportStamper: exportWalletIframeStamper,
+        exportAs: "SEED_PHRASE",
+      });
     }
 
-    return this.exportAsPrivateKey(exportWalletIframeStamper);
-  };
-
-  public getUser(): User | null {
-    return this.user ?? null;
-  }
-
-  private exportAsSeedPhrase = async (stamper: IframeStamper) => {
-    if (!this.user) {
-      throw new NotAuthenticatedError();
-    }
-
-    const { wallets } = await this.turnkeyClient.getWallets({
-      organizationId: this.user.orgId,
+    return this.exportWalletInner({
+      exportStamper: exportWalletIframeStamper,
+      exportAs: "PRIVATE_KEY",
     });
-
-    const walletAccounts = await Promise.all(
-      wallets.map(({ walletId }) =>
-        this.turnkeyClient.getWalletAccounts({
-          organizationId: this.user!.orgId,
-          walletId,
-        })
-      )
-    ).then((x) => x.flatMap((x) => x.accounts));
-
-    const walletAccount = walletAccounts.find(
-      (x) => x.address === this.user!.address
-    );
-
-    if (!walletAccount) {
-      throw new Error(
-        `Could not find wallet associated with ${this.user.address}`
-      );
-    }
-
-    const { activity } = await this.turnkeyClient.exportWallet({
-      organizationId: this.user.orgId,
-      type: "ACTIVITY_TYPE_EXPORT_WALLET",
-      timestampMs: Date.now().toString(),
-      parameters: {
-        walletId: walletAccount!.walletId,
-        targetPublicKey: stamper.publicKey()!,
-      },
-    });
-
-    const { exportBundle } = await this.pollActivityCompletion(
-      activity,
-      this.user.orgId,
-      "exportWalletResult"
-    );
-
-    const result = await stamper.injectWalletExportBundle(exportBundle);
-
-    if (!result) {
-      throw new Error("Failed to inject wallet export bundle");
-    }
-
-    return result;
-  };
-
-  private exportAsPrivateKey = async (stamper: IframeStamper) => {
-    if (!this.user) {
-      throw new NotAuthenticatedError();
-    }
-
-    const { activity } = await this.turnkeyClient.exportWalletAccount({
-      organizationId: this.user.orgId,
-      type: "ACTIVITY_TYPE_EXPORT_WALLET_ACCOUNT",
-      timestampMs: Date.now().toString(),
-      parameters: {
-        address: this.user.address,
-        targetPublicKey: stamper.publicKey()!,
-      },
-    });
-
-    const { exportBundle } = await this.pollActivityCompletion(
-      activity,
-      this.user.orgId,
-      "exportWalletAccountResult"
-    );
-
-    const result = await stamper.injectKeyExportBundle(exportBundle);
-
-    if (!result) {
-      throw new Error("Failed to inject wallet export bundle");
-    }
-
-    return result;
-  };
-
-  public addPasskey = async (options: CredentialCreationOptions) => {
-    if (!this.user) {
-      throw new NotAuthenticatedError();
-    }
-    const { attestation, challenge } = await this.getWebAuthnAttestation(
-      options
-    );
-
-    const { activity } = await this.turnkeyClient.createAuthenticators({
-      type: "ACTIVITY_TYPE_CREATE_AUTHENTICATORS_V2",
-      timestampMs: Date.now().toString(),
-      organizationId: this.user.orgId,
-      parameters: {
-        userId: this.user.userId,
-        authenticators: [
-          {
-            attestation,
-            authenticatorName: `passkey-${Date.now().toString()}`,
-            challenge: base64UrlEncode(challenge),
-          },
-        ],
-      },
-    });
-
-    const { authenticatorIds } = await this.pollActivityCompletion(
-      activity,
-      this.user.orgId,
-      "createAuthenticatorsResult"
-    );
-
-    return authenticatorIds;
-  };
-
-  public request = async <R extends SignerRoutes>(
-    route: R,
-    body: SignerBody<R>
-  ): Promise<SignerResponse<R>> => {
-    const url = this.connectionConfig.rpcUrl ?? "https://api.g.alchemy.com";
-    const basePath = "/signer";
-
-    const headers = new Headers();
-    headers.append("Content-Type", "application/json");
-    if (this.connectionConfig.apiKey) {
-      headers.append("Authorization", `Bearer ${this.connectionConfig.apiKey}`);
-    } else if (this.connectionConfig.jwt) {
-      headers.append("Authorization", `Bearer ${this.connectionConfig.jwt}`);
-    }
-
-    const response = await fetch(`${url}${basePath}${route}`, {
-      method: "POST",
-      body: JSON.stringify(body),
-      headers,
-    });
-
-    if (!response.ok) {
-      throw new Error(await response.text());
-    }
-
-    const json = await response.json();
-
-    return json as SignerResponse<R>;
   };
 
   public disconnect = async () => {
@@ -447,66 +186,18 @@ export class AlchemySignerClient {
     this.iframeStamper.clear();
   };
 
-  private pollActivityCompletion = async <
-    T extends keyof Awaited<
-      ReturnType<(typeof this.turnkeyClient)["getActivity"]>
-    >["activity"]["result"]
-  >(
-    activity: Awaited<
-      ReturnType<(typeof this.turnkeyClient)["getActivity"]>
-    >["activity"],
-    organizationId: string,
-    resultKey: T
-  ): Promise<
-    NonNullable<
-      Awaited<
-        ReturnType<(typeof this.turnkeyClient)["getActivity"]>
-      >["activity"]["result"][T]
-    >
-  > => {
-    if (activity.status === "ACTIVITY_STATUS_COMPLETED") {
-      return activity.result[resultKey]!;
-    }
-
-    const {
-      activity: { status, id, result },
-    } = await this.turnkeyClient.getActivity({
-      activityId: activity.id,
-      organizationId,
-    });
-
-    if (status === "ACTIVITY_STATUS_COMPLETED") {
-      return result[resultKey]!;
-    }
-
-    if (
-      status === "ACTIVITY_STATUS_FAILED" ||
-      status === "ACTIVITY_STATUS_REJECTED" ||
-      status === "ACTIVITY_STATUS_CONSENSUS_NEEDED"
-    ) {
-      throw new Error(
-        `Failed to get activity with with id ${id} (status: ${status})`
-      );
-    }
-
-    // TODO: add ability to configure this + add exponential backoff
-    await new Promise((resolve) => setTimeout(resolve, 500));
-
-    return this.pollActivityCompletion(activity, organizationId, resultKey);
-  };
-
   private initIframeStamper = async () => {
     if (!this.iframeStamper.publicKey()) {
       await this.iframeStamper.init();
     }
 
-    this.turnkeyClient.stamper = this.iframeStamper;
+    this.setStamper(this.iframeStamper);
 
     return this.iframeStamper.publicKey()!;
   };
 
   private initWebauthnStamper = async (user: User | undefined = this.user) => {
-    this.turnkeyClient.stamper = this.webauthnStamper;
+    this.setStamper(this.webauthnStamper);
     if (user && user.credentialId) {
       // The goal here is to allow us to cache the allowed credential, but this doesn't work with hybrid transport :(
       this.webauthnStamper.allowCredentials = [
@@ -519,7 +210,7 @@ export class AlchemySignerClient {
     }
   };
 
-  private getWebAuthnAttestation = async (
+  protected getWebAuthnAttestation = async (
     options?: CredentialCreationOptionOverrides,
     userDetails: { username: string } = {
       username: this.user?.email ?? "anonymous",
